delete partial_update_user and combining partial and full updates into one file
